Split utility functions into own file.  Utilize decorators for different subcommands.  Add subcommand for write and for get.  Eventually will add a subcommand for delete as well (requires Azure Function first).
